[Game/Zones] Simple move refactor to differentiate between logic and graphics for zones
Took 21 minutes
