+++ cockatrice/CMakeLists.txt
@@ -104,20 +104,20 @@ set(cockatrice_SOURCES
     src/game/player/player_manager.cpp
     src/game/player/player_target.cpp
     src/game/replay.cpp
-    src/game/zones/card_zone.cpp
-    src/game/zones/hand_zone.cpp
     src/game/zones/logic/card_zone_logic.cpp
     src/game/zones/logic/hand_zone_logic.cpp
     src/game/zones/logic/pile_zone_logic.cpp
     src/game/zones/logic/stack_zone_logic.cpp
     src/game/zones/logic/table_zone_logic.cpp
     src/game/zones/logic/view_zone_logic.cpp
-    src/game/zones/pile_zone.cpp
-    src/game/zones/select_zone.cpp
-    src/game/zones/stack_zone.cpp
-    src/game/zones/table_zone.cpp
-    src/game/zones/view_zone.cpp
-    src/game/zones/view_zone_widget.cpp
+    src/game_graphics/zones/card_zone.cpp
+    src/game_graphics/zones/hand_zone.cpp
+    src/game_graphics/zones/pile_zone.cpp
+    src/game_graphics/zones/select_zone.cpp
+    src/game_graphics/zones/stack_zone.cpp
+    src/game_graphics/zones/table_zone.cpp
+    src/game_graphics/zones/view_zone.cpp
+    src/game_graphics/zones/view_zone_widget.cpp
     src/game_graphics/board/abstract_graphics_item.cpp
     src/interface/card_picture_loader/card_picture_loader.cpp
     src/interface/card_picture_loader/card_picture_loader_local.cpp
