+++ cockatrice/src/game/player/player.cpp
@@ -1,5 +1,9 @@
 #include "player.h"
 
+#include "../../game_graphics/zones/hand_zone.h"
+#include "../../game_graphics/zones/pile_zone.h"
+#include "../../game_graphics/zones/stack_zone.h"
+#include "../../game_graphics/zones/table_zone.h"
 #include "../../interface/theme_manager.h"
 #include "../../interface/widgets/tabs/tab_game.h"
 #include "../board/arrow_item.h"
@@ -7,10 +11,6 @@
 #include "../board/card_list.h"
 #include "../board/counter_general.h"
 #include "../game_scene.h"
-#include "../zones/hand_zone.h"
-#include "../zones/pile_zone.h"
-#include "../zones/stack_zone.h"
-#include "../zones/table_zone.h"
 #include "player_actions.h"
 #include "player_target.h"
