+++ cockatrice/src/game/player/player_event_handler.cpp
@@ -1,10 +1,10 @@
 #include "player_event_handler.h"
 
+#include "../../game_graphics/zones/view_zone.h"
 #include "../../interface/widgets/tabs/tab_game.h"
 #include "../board/arrow_item.h"
 #include "../board/card_item.h"
 #include "../board/card_list.h"
-#include "../zones/view_zone.h"
 #include "player.h"
 #include "player_actions.h"
