+++ cockatrice/src/game/zones/logic/card_zone_logic.cpp
@@ -1,9 +1,9 @@
 #include "card_zone_logic.h"
 
+#include "../../../game_graphics/zones/view_zone.h"
 #include "../../board/card_item.h"
 #include "../../player/player.h"
 #include "../../player/player_actions.h"
-#include "../view_zone.h"
 #include "view_zone_logic.h"
 
 #include <QAction>
