+++ cockatrice/src/game_graphics/zones/card_zone.h
@@ -7,9 +7,9 @@
 #ifndef CARDZONE_H
 #define CARDZONE_H
 
-#include "../../game_graphics/board/abstract_graphics_item.h"
-#include "../../game_graphics/board/graphics_item_type.h"
-#include "logic/card_zone_logic.h"
+#include "../../game/zones/logic/card_zone_logic.h"
+#include "../board/abstract_graphics_item.h"
+#include "../board/graphics_item_type.h"
 
 #include <QLoggingCategory>
 #include <QString>
