+++ cockatrice/src/game_graphics/zones/hand_zone.cpp
@@ -1,11 +1,11 @@
 #include "hand_zone.h"
 
 #include "../../client/settings/cache_settings.h"
+#include "../../game/board/card_drag_item.h"
+#include "../../game/board/card_item.h"
+#include "../../game/player/player.h"
+#include "../../game/player/player_actions.h"
 #include "../../interface/theme_manager.h"
-#include "../board/card_drag_item.h"
-#include "../board/card_item.h"
-#include "../player/player.h"
-#include "../player/player_actions.h"
 
 #include <QPainter>
 #include <libcockatrice/protocol/pb/command_move_card.pb.h>
