+++ cockatrice/src/game_graphics/zones/pile_zone.cpp
@@ -1,10 +1,10 @@
 #include "pile_zone.h"
 
-#include "../board/card_drag_item.h"
-#include "../board/card_item.h"
-#include "../player/player.h"
-#include "../player/player_actions.h"
-#include "logic/pile_zone_logic.h"
+#include "../../game/board/card_drag_item.h"
+#include "../../game/board/card_item.h"
+#include "../../game/player/player.h"
+#include "../../game/player/player_actions.h"
+#include "../../game/zones/logic/pile_zone_logic.h"
 #include "view_zone.h"
 
 #include <QApplication>
