+++ cockatrice/src/game_graphics/zones/pile_zone.h
@@ -7,8 +7,8 @@
 #ifndef PILEZONE_H
 #define PILEZONE_H
 
+#include "../../game/zones/logic/pile_zone_logic.h"
 #include "card_zone.h"
-#include "logic/pile_zone_logic.h"
 
 /**
  * A CardZone where the cards are in a single pile instead of being laid out.
