+++ cockatrice/src/game_graphics/zones/stack_zone.h
@@ -1,35 +1,32 @@
 /**
- * @file hand_zone.h
+ * @file stack_zone.h
  * @ingroup GameGraphicsZones
- * @brief Graphical zone for the player's hand, supporting horizontal and vertical layouts.
+ * @brief Graphical zone for the stack, displaying cards in a vertical pile.
  */
 
-#ifndef HANDZONE_H
-#define HANDZONE_H
+#ifndef STACKZONE_H
+#define STACKZONE_H
 
-#include "logic/hand_zone_logic.h"
+#include "../../game/zones/logic/stack_zone_logic.h"
 #include "select_zone.h"
 
-class HandZone : public SelectZone
+class StackZone : public SelectZone
 {
     Q_OBJECT
 private:
-    qreal width = 0.0;
     qreal zoneHeight;
 private slots:
     void updateBg();
-public slots:
-    void updateOrientation();
 
 public:
-    HandZone(HandZoneLogic *_logic, int _zoneHeight, QGraphicsItem *parent = nullptr);
+    StackZone(StackZoneLogic *_logic, int _zoneHeight, QGraphicsItem *parent);
+    /// @brief Resizes the stack zone height, e.g. when sharing vertical space with the command zone.
+    void setHeight(qreal newHeight);
     void
     handleDropEvent(const QList<CardDragItem *> &dragItems, CardZoneLogic *startZone, const QPoint &dropPoint) override;
     QRectF boundingRect() const override;
     void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget) override;
     void reorganizeCards() override;
-    void sortHand(const QList<CardList::SortOption> &options);
-    void setWidth(qreal _width);
 };
 
 #endif
