+++ cockatrice/src/game_graphics/zones/table_zone.cpp
@@ -1,14 +1,14 @@
 #include "table_zone.h"
 
 #include "../../client/settings/cache_settings.h"
+#include "../../game/board/arrow_item.h"
+#include "../../game/board/card_drag_item.h"
+#include "../../game/board/card_item.h"
+#include "../../game/player/player.h"
+#include "../../game/player/player_actions.h"
+#include "../../game/z_values.h"
+#include "../../game/zones/logic/table_zone_logic.h"
 #include "../../interface/theme_manager.h"
-#include "../board/arrow_item.h"
-#include "../board/card_drag_item.h"
-#include "../board/card_item.h"
-#include "../player/player.h"
-#include "../player/player_actions.h"
-#include "../z_values.h"
-#include "logic/table_zone_logic.h"
 
 #include <QGraphicsScene>
 #include <QPainter>
