+++ cockatrice/src/game_graphics/zones/table_zone.h
@@ -7,8 +7,8 @@
 #ifndef TABLEZONE_H
 #define TABLEZONE_H
 
-#include "../board/abstract_card_item.h"
-#include "logic/table_zone_logic.h"
+#include "../../game/board/abstract_card_item.h"
+#include "../../game/zones/logic/table_zone_logic.h"
 #include "select_zone.h"
 
 /*
